Add models and forms to project, react with database.
The website is now functional with the ability to create a choir when a person signs up. The are able to create a group login.
There still could be a nice tab that lets the user check out the group login name. The hide random words is not functional yet.